+++ thali/install/install.js
@@ -1,4 +1,5 @@
 'use strict';
+var childProcessExecPromise = require('./utils.js').childProcessExecPromise;
 var exec = require('child_process').exec;
 var spawn = require('child_process').spawn;
 var path = require('path');
@@ -13,25 +14,6 @@ var FILE_NOT_FOUND = 'ENOENT';
 // we will copy the Cordova plugin from a sibling Thali_CordovaPlugin
 // project to this Cordova project.
 var MAGIC_DIRECTORY_NAME_FOR_LOCAL_DEPLOYMENT = 'localdev';
-
-// I tried child-process-promise but it failed without errors and I just don't
-// have time to fight with it right now.
-function childProcessExecPromise(command, currentWorkingDirectory) {
-  return new Promise(function (resolve, reject) {
-    exec(command, { cwd: currentWorkingDirectory },
-      function (error, stdout, stderr) {
-        if (error) {
-          reject(error);
-          return;
-        }
-        // Log output even if command doesn't exit with an error,
-        // because otherwise useful debugging information might get lost.
-        if (stdout) { console.log(stdout); }
-        if (stderr) { console.log(stderr); }
-        resolve();
-      });
-  });
-}
 
 // Unfortunately the obvious library, request-promise, doesn't handle streams
 // well so it would take the multi-megabyte ZIP response file and turn it into
