+++ thali/package.json
@@ -16,8 +16,7 @@
     "node-uuid": "^1.4.7",
     "request": "^2.64.0",
     "urlsafe-base64": "^1.0.0",
-    "winston": "^2.2.0",
-    "xcode": "^0.8.9"
+    "winston": "^2.2.0"
   },
   "devDependencies": {
     "jsdoc": "^3.3.3"
